+++ beem/__init__.py
@@ -1,6 +1,7 @@
 """ beem."""
 from .steem import Steem
 from .hive import Hive
+from .blurt import Blurt
 from .version import version as __version__
 __all__ = [
     "steem",
@@ -8,6 +9,7 @@ __all__ = [
     "amount",
     "asset",
     "block",
+    "blurt",
     "blockchain",
     "blockchaininstance",
     "market",
@@ -25,6 +27,5 @@ __all__ = [
     "nodelist",
     "imageuploader",
     "snapshot",
-    "steemconnect",
     "hivesigner"
 ]